requirements.txt
streamlit>=1.20
PyYAML
python-dotenv
#nltk
#transformers>=4.0.0
#torch
